requirements.txt
notion-client>=2.0.0
playwright>=1.40.0
openai>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
loguru>=0.7.0
APScheduler>=3.10.0
PyYAML>=6.0
aiohttp>=3.9.0

